homogenization of specifis urls localized
