+++ apps/ledger-live-desktop/src/renderer/components/Onboarding/Screens/Tutorial/index.tsx
@@ -48,7 +48,7 @@ import { RecoveryHowTo3 } from "~/renderer/components/Onboarding/Screens/Tutoria
 import { UseRecoverySheet } from "~/renderer/components/Onboarding/Screens/Tutorial/screens/UseRecoverySheet";
 import { openURL } from "~/renderer/linking";
 import { QuizzPopin } from "~/renderer/modals/OnboardingQuizz/OnboardingQuizzModal";
-import { useDynamicUrl } from "~/renderer/terms";
+import { useDynamicUrl } from "~/config/urls-localized";
 import RecoveryWarning from "../../Help/RecoveryWarning";
 import { UseCase } from "../../index";
