+++ apps/ledger-live-desktop/src/renderer/components/Web3AppWebview/NetworkError.tsx
@@ -4,7 +4,7 @@ import { useTranslation } from "react-i18next";
 import ErrorNoBorder from "~/renderer/icons/ErrorNoBorder";
 import { openURL } from "~/renderer/linking";
 import ExportLogsButton from "~/renderer/components/ExportLogsButton";
-import { useDynamicUrl } from "~/renderer/terms";
+import { useDynamicUrl } from "~/config/urls-localized";
 
 export const NetworkErrorScreen = ({ refresh }: { refresh: () => void }) => {
   const { t } = useTranslation();
