+++ apps/ledger-live-desktop/src/renderer/families/cosmos/StakeBanner.tsx
@@ -10,9 +10,11 @@ import { openModal } from "~/renderer/actions/modals";
 import { useDispatch } from "react-redux";
 import { track } from "~/renderer/analytics/segment";
 import { stakeDefaultTrack } from "~/renderer/screens/stake/constants";
+import { StakingCoin, useStakingUrl } from "~/config/urls-localized";
 
 const StakeBanner: React.FC<{ account: CosmosAccount }> = ({ account }) => {
   const dispatch = useDispatch();
+  const stakingUrl = useStakingUrl(StakingCoin.cosmos);
   const { t } = useTranslation();
   const stakeAccountBanner = useFeature("stakeAccountBanner");
   const stakeAccountBannerParams: StakeAccountBannerParams | null =
@@ -73,7 +75,7 @@ const StakeBanner: React.FC<{ account: CosmosAccount }> = ({ account }) => {
       onClick={onClick}
       display={display}
       linkText={t("account.banner.delegation.linkText")}
-      linkUrl={"https://www.ledger.com/staking/ledger-node/cosmos"}
+      linkUrl={stakingUrl}
     />
   );
 };
