+++ apps/ledger-live-desktop/src/renderer/families/tezos/DelegateFlowModal/steps/StepStarter.tsx
@@ -10,7 +10,7 @@ import Button from "~/renderer/components/Button";
 import Box from "~/renderer/components/Box";
 import LinkWithExternalIcon from "~/renderer/components/LinkWithExternalIcon";
 import { StepProps } from "../types";
-import { StakingCoin, useStakingUrl } from "~/config/urls-staking";
+import { StakingCoin, useStakingUrl } from "~/config/urls-localized";
 const Row = styled(Box).attrs(p => ({
   horizontal: true,
   justifyContent: "flex-start",
