+++ apps/ledger-live-desktop/src/renderer/screens/dashboard/EmptyStateInstalledApps.tsx
@@ -9,7 +9,7 @@ import Text from "~/renderer/components/Text";
 import LinkHelp from "~/renderer/components/LinkHelp";
 import { openURL } from "~/renderer/linking";
 import { DefaultTheme, withTheme } from "styled-components";
-import { useDynamicUrl } from "~/renderer/terms";
+import { useDynamicUrl } from "~/config/urls-localized";
 
 const EmptyStateInstalledApps = ({ theme }: { theme: DefaultTheme }) => {
   const { push } = useHistory();
